requirements.txt
wheel
pymysql
emoji
requests
pillow
datetime
wavelink
ffmpeg
python-dotenv
pynacl
aiohttp
